requirements.txt
python-telegram-bot==21.5
motor==3.3.2
pymongo==4.6.3
langgraph==0.1.19
langchain==0.2.16
langchain-community==0.2.16
fastapi==0.111.1
uvicorn==0.29.0
pydantic==2.7.4
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.0
jinja2==3.1.4
python-multipart==0.0.9
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
